+++ grizly/sources/__init__.py
@@ -1,4 +1,4 @@
 # from .rdbms.aurora import AuroraDB
-from .rdbms.sfdc import SFDB, sfdb
+from .rdbms.sfdc import SFDB
 from .filesystem.local import LocalFS
 from .filesystem.old_s3 import S3