+++ tests/test_s3.py
@@ -80,19 +80,19 @@ def test_to_rds():
 
     qf.to_parquet(path_parquet)
     s3_parquet.from_file(keep_file=False)
-    s3_parquet.to_rds(table=table_parquet, schema=schema, if_exists="replace")
+    s3_parquet.to_rds(dsn="redshift_acoe", table=table_parquet, schema=schema, if_exists="replace")
     qf_parquet = QFrame(dsn="redshift_acoe").from_table(table=table_parquet, schema=schema)
     assert len(qf_parquet) == 30
 
     qf.to_csv(path_csv)
     s3_csv.from_file(keep_file=False)
-    s3_csv.to_rds(table=table_csv, schema=schema, if_exists="replace")
+    s3_csv.to_rds(dsn="redshift_acoe", table=table_csv, schema=schema, if_exists="replace")
     qf_csv = QFrame(dsn="redshift_acoe").from_table(table=table_csv, schema=schema)
     assert len(qf_csv) == 30
 
     qf.to_parquet(path_parquet)
     s3_parquet.from_file(keep_file=False)
-    s3_parquet.to_rds(table=table_parquet, schema=schema, if_exists="append")
+    s3_parquet.to_rds(dsn="redshift_acoe", table=table_parquet, schema=schema, if_exists="append")
     assert len(qf_parquet) == 60
 
     qf.rearrange(
@@ -112,7 +112,7 @@ def test_to_rds():
     )
     qf.to_csv(path_csv)
     s3_csv.from_file(keep_file=False)
-    s3_csv.to_rds(table=table_csv, schema=schema, if_exists="append")
+    s3_csv.to_rds(dsn="redshift_acoe", table=table_csv, schema=schema, if_exists="append")
     assert len(qf_csv) == 60
 
     qf_csv.distinct()